requirements.txt
deepgram-sdk==3.2.1
edge-tts==6.1.10
langchain==0.1.12
langchain_core==0.1.32
langchain_groq==0.0.1
langchain_openai==0.0.8
langchain-community==0.0.28
openai-whisper==20231117
PyAudio==0.2.14
python-dotenv==1.0.1
sounddevice==0.4.6
torch==2.2.1
torchaudio==2.2.1
TTS==0.22.0